move session auth headers to request interceptor
